+++ main.py
@@ -7,6 +7,9 @@ from signal_generators import SignalGenerator
 from base_strategy import MACDStateMachine
 from binance_account import BinanceAccount
 from models.run_setting import RunSetting
+from models.run import Run
+
+import time
 
 ## API_KEY
 api_key = config.API_KEY
@@ -19,12 +22,26 @@ timestamp = startTime.timestamp()
 startTimeStr = startTime.strftime("%m/%d/%Y - %H:%M:%S")
 print(startTimeStr)
 
-# get run settings
-runSetting = RunSetting.query(limit=1, slug="default_bnbbusd")[0]
+# search for any 
+inturruptedRun = None
+runs = Run.query(limit=1, end=None)
+if len(runs) > 0: 
+    inturruptedRun = runs[0]
+
+## -------- IMPORTANT VARS !
+quoteFundAmount = 40 # TEMP - how much the bot is going to trade
+
+runSetting = None # general settings of a run
+
+# start fresh
+if inturruptedRun is None:
+    # get run settings
+    runSetting = RunSetting.query(limit=1, slug="default_bnbbusd")[0]
+else:
+    runSettingId = inturruptedRun.runSettingId
+    runSetting = RunSetting.get(runSettingId)
 
 ## run 
-cc = runSetting.websocketSymbol
-
 interval, lookback = None, None
 if runSetting.tradeInterval == "ONE_MINUTE":
     interval, lookback = kline_lookback_config.ONE_MINUTE
@@ -33,6 +50,8 @@ elif runSetting.tradeInterval == "FIVE_MINUTE":
 elif runSetting.tradeInterval == "FIFTEEN_MINUTE":
     interval, lookback = kline_lookback_config.FIFTEEN_MINUTE
 
+
+cc = runSetting.websocketSymbol
 socket = f"wss://stream.binance.com:9443/ws/{cc}@kline_{interval}"
 
 # TODO - gat from account
@@ -40,7 +59,7 @@ client = Client(config.API_KEY, config.API_SECRET)
 
 symbol = runSetting.symbol
 
-indicators = runSetting.indicators
+indicators = runSetting.indicators # EMAS MACD
 
 # signal generator
 signal = SignalGenerator(client, symbol, lookback, interval)
@@ -48,8 +67,21 @@ signal = SignalGenerator(client, symbol, lookback, interval)
 #BinanceAccount
 account = BinanceAccount(client, symbol)
 
-# bot state
-botState = MACDStateMachine(account, symbol, riskTolerancePercentage=0.5, quoteFundAmount=15, isTestNet=False)
+'''
+riskTolerancePercentage=runSetting.riskTolerancePercentage
+testNet = runSetting.testNet
+'''
+# start fresh
+if inturruptedRun is None:
+    botState = MACDStateMachine(account, symbol, riskTolerancePercentage=0.5, quoteFundAmount=quoteFundAmount, isTestNet=False)
+else:
+    botState = MACDStateMachine.resumeRun(account, inturruptedRun)
+    print("PICKED UP RUN")
+
+print(botState.getState())
+
+# botState.quoteFundAmount=20
+# botState.stopLossPrice = 299.0
 
 # set up ws class
 ws = websocket.WebSocketApp(socket, on_message=signal.getOnMessage(botState, indicators=indicators), on_close=signal.getOnClose())